Bot

Description

Bot module containing al functions used for running an IRC bot

Documentation

type Net = ReaderT Config IO

The Net monad, a wrapper over IO, carrying the bot's immutable config.

data Config

Datatype containing bots configutation

socket :: Config -> Handle

returns Handle to the socket bot is connected to

connect :: IO Config

Connect to the server and return the initial bot state

run :: Net ()

Sets bot parameters and listens in a loop for incoming messages