Weather

Description

Weather module used for getting current weather information for Krakow

Documentation

getWeatherKrk :: IO String

Returns string containing information about current weather in Krakow. Weather is downloaded from openWeathermap API. Because it's downloaded from the internet it's wrapped in IO.

jsonURL :: String

OpenWeatherMap API URL for Krakow

getJSON :: IO ByteString

Downloads json from jsonURL

getTemp :: Maybe Weather -> String

Extracts temperature information from result of decoding JSON

getPressure :: Maybe Weather -> String

Extracts pressure information from result of decoding JSON

getHumidity :: Maybe Weather -> String

Extracts humidity information from result of decoding JSON

data Weather

Data type containing information about current weather